Cherrypick stab to dev 03 (#18389)
* Script canvas logger can now identify graph and entity names (#17912)

* Script canvas logger can now identify graph and entity names

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>

* Init remote tools later to prevent timing issues

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>

---------

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>
Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Further fixes for the ScriptCanvas debugging facilities (#18145)

* Keep scriptcanvas and compiled files entity UIDs in sync and fix node focus on trace click

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>

* Add notes for future fix

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>

* Remove useless check

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>

---------

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>
Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fixes a crash if you enable Script canvas developer Gem (#18328)

Fix issue https://github.com/o3de/o3de/issues/17572

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fixes a crash when you try to remove elements from unordered set. (#18352)

Fixes issue https://github.com/o3de/o3de/issues/18348

The function that it depended on was simply not implemented.
While it is not ideal for it to walk the memory linearly when
removing an element from a set, it's still worth implementing for
cases where the set is being used in a way that relies on indices,
which is the case for the UI.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fix crash after leaving picking mode (#18366)

fixes issue https://github.com/o3de/o3de/issues/18365

The hook to register and unregister actions happens once during startup,
and is not meant to have any connection to the lifetime of the objects.

The crash was happening because the EditorTransformComponentSelection
object was being destroyed and recreated every time you enter pick mode,
but its action registration captured its 'this' pointer.

usually, such action registration would be placed inside some kind of long
lived singleton "manager" object, aka a system component, but in this case,
it was put in an ephemeral object.  note that this is still a singleton
object, there is only ever one of it, and it is essentially the "manager",
its just special in that sometimes, it does not exist.

to fix this with as little "splash damage" as possible, I kept most
of the code exactly as-is, without modifying any other classes,
but I made the action registration happen in a static function with
no "this" pointer, which means the compiler could find every illegal
instance.

I added a simple AZ::Environment based singleton pattern in there to get
the current instance when available.  This is also an attempt to keep
the "Blast radius" of this change as small as possible, as this is a
well-used pattern and bypasses any chance of there being weird
dll-related issues.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Revert "Fix hyper-sensitve camera rotation remote clients (#18354)" (#18368)

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Update O3DE Splash Screen background (#18372)

Signed-off-by: Steve Pham <82231385+spham-amzn@users.noreply.github.com>
Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: guillaume-haerinck <guillaume.haerinck@outlook.com>
Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
Signed-off-by: Steve Pham <82231385+spham-amzn@users.noreply.github.com>
Co-authored-by: Guillaume Haerinck <guillaume.haerinck@outlook.com>
Co-authored-by: Steve Pham <82231385+spham-amzn@users.noreply.github.com>